requirements.txt
adversarial-robustness-toolbox==1.7.2
# ipykernel==6.4.1
ipywidgets==7.6.5
imutils==0.5.4
kaleido==0.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lime==0.2.0.1
matplotlib==3.3.4
nbformat==5.1.3
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==8.3.1
plotly==5.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
shap==0.39.0
shortuuid==1.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tqdm==4.62.0
traitlets==4.3.3
typing-extensions==3.7.4.3
